+++ check.js
@@ -4,7 +4,8 @@ var server = new index.WebServer({
   port: 3000,
   acceptonlyget: true,
   notfound: "This page shows that the url requested was not found on this server!",
-  req405error: "This page shows that you sent a request with incorrect method!"
+  req405error: "This page shows that you sent a request with incorrect method!",
+  serveindex: true
 });
 
 var filename = "./check.html";