+++ index.js
@@ -1,4 +1,6 @@
 
+'use strict'
+
 /*
     Copyright (c) 2021 chasyumen
     MIT Licensed