+++ src/WebServer.js
@@ -9,13 +9,14 @@ const serveIndex = require("./serve-index-1.9.1-modded/index.js")
  * @typedef WebServerOptions
  * @type {object}
  * @property {object} [options]
+ * @property {Number} [options.port=3000] - Port to host the page.
  * @property {string} [options.directory=./public/] - The directory to load the file. 
+ * @property {boolean} [options.serveindex=false] - If it's true, return the file list of the directory when the directory was requested. (If index.html exists in the directory, it will be sent preferentially)
  * @property {boolean} [options.acceptonlyget=true] - Only accepts get request.
- * @property {string} [options.errordocument._404=`${__dirname}/assets/def_pages/404.html`] - Html file path to respond on the request methods other than GET.
- * @property {string} [options.errordocument._405=`${__dirname}/assets/def_pages/405.html`] - Html file path to respond on not found.
  * @property {boolean} [options.useindexhtml=true] - If it's true, returns ./index.html file when requested directory.
  * @property {string} [options.rootfile=/index.html] - You can specify a file to load as a top page. (Please specify files in the published directory.)
- * @property {Number} [options.port=3000] - Port to host the page.
+ * @property {string} [options.errordocument._404=`${__dirname}/assets/def_pages/404.html`] - Html file path to respond on the request methods other than GET.
+ * @property {string} [options.errordocument._405=`${__dirname}/assets/def_pages/405.html`] - Html file path to respond on not found.
  */
 
 /**
@@ -55,7 +56,7 @@ class WebServer {
       }
       return options;
     }
-    
+
   }
 
   /**
@@ -69,6 +70,7 @@ class WebServer {
       if (!opt) {
         var options = {
           directory: "./public",
+          serveindex: false,
           //notfound: `<!DOCTYPE html>\n<html>\n<head>\n<title>404 Not Found</title>\n</head>\n<body>\n<center>\n<h1>404 Not Found</h1>\n</center>\n<hr>\n<center>Node.js lite-web-server</center>\n</body>\n</html>`,
           port: 3000,
           acceptonlyget: true,
@@ -104,11 +106,17 @@ class WebServer {
         } else {
           var useindexhtml = true;
         }
+        if (opt.serveindex === true) {
+          var serveindex = true;
+        } else {
+          var serveindex = false;
+        }
         if (!opt.errordocument) {
           opt.errordocument = {};
         }
         var options = {
           directory: dir || "./public",
+          serveindex: serveindex,
           //notfound: opt.notfound || `<!DOCTYPE html>\n<html>\n<head>\n<title>404 Not Found</title>\n</head>\n<body>\n<center>\n<h1>404 Not Found</h1>\n</center>\n<hr>\n<center>Node.js lite-web-server</center>\n</body>\n</html>`,
           port: opt.port || 3000,
           acceptonlyget: acceptonlyget,
@@ -130,7 +138,7 @@ class WebServer {
       }
       try {
         var httpserver = http.createServer(async function (req, res) {
-          if (!(req.method.toUpperCase() == "GET")&&options.acceptonlyget == true) {
+          if (!(req.method.toUpperCase() == "GET") && options.acceptonlyget == true) {
             try {
               var read_file = await fs.readFileSync(options.errordocument._405).toString();
               res.writeHead(405, { "Content-Type": "text/html" });
@@ -144,7 +152,7 @@ class WebServer {
           //console.log(url);
           if (url == "/") {
             var filedir = `${options.directory}${options.rootfile}`;
-          } else if (url.endsWith("/")&&options.useindexhtml) {
+          } else if (url.endsWith("/") && options.useindexhtml) {
             var filedir = `${options.directory}${url}index.html`;
           } else {
             var filedir = `${options.directory}${url}`;
@@ -157,30 +165,51 @@ class WebServer {
           } catch (error) {
             try {
               try {
-                var _loaddirurl = options.directory+"/";
+                var _loaddirurl = options.directory + "/";
                 if (url.endsWith("/")) {
                   var __loaddirurl = _loaddirurl
+                  try {
+                    //console.log(_loaddirurl+url.slice(1))
+                    fs.readdirSync(_loaddirurl + url.slice(1))
+                  } catch (error) {
+                    throw new Error(error)
+                  }
                 } else {
                   //console.log(req)
-                  res.writeHead(302, {location: `${req.url}/`,});
-                  res.end();
-                  return;
+                  try {
+                    //console.log(_loaddirurl+url.slice(1))
+                    fs.readdirSync(_loaddirurl + url.slice(1))
+                    res.writeHead(302, { location: `${req.url}/`, });
+                    res.end();
+                    return;
+                  } catch (error) {
+                    throw new Error(error)
+                  }
+
                   //var loaddirurl = _loaddirurl + "/"
                 }
                 var loaddirurl = `${__loaddirurl}`
                 //console.log(loaddirurl)
-                var returns = serveIndex(loaddirurl, {icons: true, view: "details"})(req, res);
+                try {
+                  if (options.serveindex === false) {
+                    var read_file = await fs.readFileSync(options.errordocument._404).toString();
+                    var file = read_file.replace(/<!--\${404URL}-->/g, `${url}`);
+                    res.writeHead(404, { "Content-Type": "text/html" });
+                    res.end(file);
+                    return;
+                  }
+                  var serveindex = serveIndex(loaddirurl, { icons: true, view: "details" });
+                  serveindex(req, res);
+                } catch (error) {
+                  throw new Error(error)
+                }
+
               } catch (error) {
-                var returns = false;
-              }
-              //console.log(returns)
-              if (returns === false) {
                 var read_file = await fs.readFileSync(options.errordocument._404).toString();
                 var file = read_file.replace(/<!--\${404URL}-->/g, `${url}`);
                 res.writeHead(404, { "Content-Type": "text/html" });
                 res.end(file);
               }
-              
             } catch (error) {
               res.writeHead(500, { "Content-Type": "text/html" });
               res.end("<center><h1>Internal Server Error</h1></center>");
@@ -197,7 +226,7 @@ class WebServer {
         reject(new Error(error));
       }
     });
-  } 
+  }
 }
 
 module.exports = WebServer