+++ src/serve-index-1.9.1-modded/escape-html-1.0.3.js
@@ -30,7 +30,7 @@ module.exports = escapeHtml
  * @public
  */
 
-function escapeHtml (string) {
+function escapeHtml(string) {
   var str = '' + string
   var match = matchHtmlRegExp.exec(str)