+++ check.js
@@ -1,12 +1,13 @@
-var index = require("./lib/index.js"); //this file loads ./lib/index.js
+var index = require("./src/index.js"); //this file loads ./lib/index.js
 var server = new index.WebServer({
   dir: "./",
   port: 3000,
   acceptonlyget: true,
   notfound: "This page shows that the url requested was not found on this server!",
-  req405error: "This page shows that you sent a request with incorrect method!"
+  req405error: "This page shows that you sent a request with incorrect method!",
+  serveindex: true
 });
 
-var filename = "./check.html"
-var filetype = index.GetFileType(filename)
+var filename = "./check.html";
+var filetype = index.GetFileType(filename);