+++ index.js
@@ -1,7 +1,10 @@
 
+'use strict'
+
 /*
+    lite-web-server
     Copyright (c) 2021 chasyumen
     MIT Licensed
 */
 
-module.exports = require('./lib');
+module.exports = require('./src');