+++ src/index.js
@@ -1,0 +1,9 @@
+'use strict'
+
+const GetFileType = require("./GetFileType/index.js");
+const WebServer = require("./WebServer.js");
+
+module.exports = {
+  WebServer: WebServer,
+  GetFileType: GetFileType
+}