+++ src/mime-types/type-db.js
@@ -1,0 +1,8 @@
+'use strict'
+
+/*
+  Using mine-db files.
+
+ */
+
+module.exports = require("./mime-db.1.50.0.min.json")